+++ mayan/apps/authentication/links.py
@@ -7,8 +7,11 @@ from navigation import Link
 from .icons import icon_logout, icon_password_change
 
 
-def has_usable_password(context):
-    return context['request'].user.has_usable_password
+def has_usable_password_and_can_change_password(context):
+    return (
+        context['request'].user.has_usable_password and
+        not context['request'].user.user_options.block_password_change
+    )
 
 
 link_logout = Link(
@@ -16,6 +19,7 @@ link_logout = Link(
     text=_('Logout'), view='authentication:logout_view'
 )
 link_password_change = Link(
-    condition=has_usable_password, icon_class=icon_password_change,
-    text=_('Change password'), view='authentication:password_change_view'
+    condition=has_usable_password_and_can_change_password,
+    icon_class=icon_password_change, text=_('Change password'),
+    view='authentication:password_change_view'
 )