+++ mayan/apps/user_management/forms.py
@@ -6,5 +6,7 @@ from django.contrib.auth import get_user_model
 
 class UserForm(forms.ModelForm):
     class Meta:
+        fields = (
+            'username', 'first_name', 'last_name', 'email', 'is_active',
+        )
         model = get_user_model()
-        fields = ('username', 'first_name', 'last_name', 'email', 'is_active',)